Fix inserting a new line
When inserting a new line, it looks like you CANNOT insert it with indentation by specifying the "character" of the location. The character has to be 0, because there's no text on the line yet. So if you want it indented, you have to add spaces at the beginning of the inserted text